Move product tests to the right test suite
